\name{seatdist-package}
\alias{seatdist-package}
\alias{seatdist}
\docType{package}
\title{
    \packageTitle{seatdist}
}
\description{
    \packageDescription{seatdist}
}
\details{
    The DESCRIPTION file:
    \packageDESCRIPTION{seatdist}
    \packageIndices{seatdist}
}
\author{
    \packageAuthor{seatdist}

    Maintainer: \packageMaintainer{seatdist}
}
\references{
    Medzihorsky, J. 2019. Understanding the D'Hondt Method
}
%% \keyword{ package }
%% \seealso{}
%% \examples{}
