\name{seatdist-package}
\alias{seatdist-package}
\alias{seatdist}
\docType{package}
\title{
    seatdist: Seat Apportionment and Disproportionality Measurement
}
\description{
    Functions for seat apportionment and
    measurement of apportionment disproportionality.
}
\details{
    The DESCRIPTION file:
    \packageDESCRIPTION{seatdist}
    \packageIndices{seatdist}
}
\author{
    Juraj Medzihorsky

    Maintainer: Juraj Medzihorsky <juraj.medzihorsky@gmail.com>
}
\references{
    Medzihorsky, J. 2019. Rethinking the D'Hondt Method. PRX 1 (1): 41-55. 
}
%% \keyword{ package }
%% \seealso{}
%% \examples{}
